Don't use getAll() to implement contains() and get() in Index manager for performance reasons.
When an index has a lot of values associated to a key (e.g. the inverse node index) getAll() will need return a large map which might be constructured on runtime. Using this huge map for more trivial tasks is quite inefficent, so implement these functions directly.

P.S FLORIAN STOP WORKING, YOU ARE ON HOLIDAY!