Simplify the API by not allowing generic types as arguments of functions for redundant cross-references in objects.
These cross-references (e.g. the setGraph(...) function on nodes) introduce a dependency on a generic type to the class and makes it impossible for sub-classes to not implement the function for the super-type. While it is possible to check for the run-time type and throw an exception checking the correctness of the generic type is much more difficult and easy to implement wrong. These cross-reference functions are all redundant in their functionality as thus can be removed easily.