Fix #130 not by using a separate class for notifying corpus graphs.
The previous attempt to fix the issue was not complete and not all functions that needed notification where implemented. The new fix uses a default GraphNotifierImpl as delegate again, but makes sure the special indexes for Salt-Types (which are not part of the salt.graph model) are added to the index manager and used when new relations/nodes are added to the corpus graph.